fix: make pageNumber and offset computed from query params
